BUG: using float for PET to ensure the right dynamic range -- do we need to go to double?
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=13778
git-svn-id: http://svn.slicer.org/Slicer4/trunk@13778 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Applications/CLI/DicomToNrrdConverter/DicomToNrrdConverter.cxx
@@ -416,7 +416,7 @@ int main(int argc, char* argv[])
   if (  modality.find("PT") != std::string::npos 
         || modality.find("ST") != std::string::npos )
     {
-      typedef itk::Image<unsigned short, 3> USVolumeType;
+      typedef itk::Image<float, 3> USVolumeType;
       itk::ImageSeriesReader<USVolumeType>::Pointer seriesReader = 
         itk::ImageSeriesReader<USVolumeType>::New();
       seriesReader->SetFileNames( filenamesInSeries );
